job-finder/requirements.txt
# Software Developer Job Finder System Requirements
# Compatible with Python 3.7+

# Core Framework
fastapi>=0.68.0,<0.100.0
uvicorn>=0.15.0,<0.25.0
celery>=5.0.0,<6.0.0
redis>=4.0.0,<5.0.0
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<1.12.0

# Web Automation & Anti-Detection
playwright>=1.20.0,<1.40.0
selenium>=4.0.0,<4.15.0
fake-useragent>=1.0.0,<1.5.0
requests-html>=0.10.0,<0.11.0

# AI & Text Processing
openai>=0.27.0,<1.0.0
PyPDF2>=2.0.0,<3.0.0
python-docx>=0.8.0,<1.0.0
Pillow>=8.0.0,<10.0.0

# Data Processing & Analysis
pandas>=1.3.0,<2.0.0
numpy>=1.20.0,<1.25.0
beautifulsoup4>=4.9.0,<4.13.0
lxml>=4.6.0,<4.10.0

# Database & Caching
psycopg2-binary>=2.8.0,<3.0.0

# Monitoring & Logging
loguru>=0.6.0,<0.8.0

# Utilities
python-dotenv>=0.19.0,<1.0.0
pydantic>=1.8.0,<2.0.0
httpx>=0.24.0,<0.26.0
aiofiles>=0.7.0,<23.0.0
schedule>=1.1.0,<1.3.0

# Development
pytest>=6.0.0,<8.0.0
